ChatBotLibraryKMUTNB-main/Chatbot_Library-main/requirements.txt
openai==1.35.3
python-dotenv==1.0.1
scikit-learn==1.5.0
scipy==1.13.1
pythainlp==5.0.4
tiktoken==0.7.0
Flask==3.0.3
pandas==2.2.2
openpyxl==3.1.4
google-generativeai
mysql-connector-python==9.0.0
streamlit
gspread
oauth2client
pprintpp
python-docx==1.1.0
google-generativeai>=0.7.0
pypdf>=4.2.0
pdfminer.six>=20231228
pdfplumber>=0.11.0
python-docx>=1.1.0
